requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
cachetools==7.0.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cryptography==42.0.8
cycler==0.12.1
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.18.1
flwr==1.9.0
fonttools==4.43.1
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.64.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.4
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.14.0
iterators==0.0.2
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jiter==0.13.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
kiwisolver==1.4.5
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
MouseInfo==0.1.3
msgpack==1.0.8
narwhals==2.18.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
neo4j==6.1.0
nest-asyncio==1.5.7
numpy==1.25.2
oauth2client==4.1.3
obd==0.7.2
openai==2.29.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
pdfminer.six==20231228
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
Pint==0.20.1
platformdirs==3.10.0
pluggy==1.0.0
ply==3.11
prompt-toolkit==3.0.39
protobuf==4.25.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==23.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyautomate==0.1.3
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.16.1
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
PySocks==1.7.1
pytest==7.2.0
python-can==4.3.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.2.2
pytweening==1.0.7
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rich==13.7.1
rpds-py==0.12.0
rsa==4.9
rubicon-objc==0.4.7
scipy==1.11.3
seaborn==0.13.0
selenium==4.17.2
Send2Trash==1.8.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.3
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
streamlit==1.55.0
tenacity==9.1.4
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.3.3
tqdm==4.67.3
traitlets==5.9.0
trio==0.24.0
trio-websocket==0.11.1
typer==0.9.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==3.0.2
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.1.9
